+++ src/main/java/gemini/myownradio/tools/io/NoiseInputStream.java
@@ -8,19 +8,19 @@ import java.io.InputStream;
  */
 public class NoiseInputStream extends InputStream {
 
-    private Integer length;
+    private Long length;
 
     public NoiseInputStream() {
         this(null);
     }
 
-    public NoiseInputStream(Integer length) {
+    public NoiseInputStream(Long length) {
         this.length = length;
     }
 
     @Override
     public int read() throws IOException {
-        if (length != null && length-- <= 0) return -1;
+        if (length != null && length-- <= 0L) return -1;
         return (int) (Math.random() * 255);
     }
 
@@ -32,7 +32,7 @@ public class NoiseInputStream extends InputStream {
     @Override
     public int read(byte[] b, int off, int len) throws IOException {
         if (length != null && length-- <= 0) return -1;
-        if (length != null && length < len) len = length;
+        if (length != null && length < len) len = length.intValue();
 
         for (int i = off; i < len; i++) {
             b[i] = (byte) (Math.random() * 255);
